improve speed with a hash date <-> file-position
This commit improves speed on large log files. A hash mapping
dates and corresponding file positions in the log file is created
before servlet starts.
Then instead of re-reading the whole file, a (file-position) call is performed
in order to "jump" to the closest position of the wanted date.

Also add a thread to update the hash if needed, when the log-file is modified.